Deleted handler_test_common.go (only used in handler_test.go)
